Summer Missions Component
Adds a new component to pull Summer Missions via API. This component delivers search results for trips and individual trip details from a single page based on the URL Parameters.

I also added a new .js file that will allow us to pull out any page parameter – either a single param or all params. This is added to the htmlbase clientlibs.

Finally I limited the campus-locator.js to only call the api when the component-specific class .locator is present on the page.