Bug fixes - Large Curate Tile and Main Nav for Oracle
Fixes content.html for Large Curated Tiles where content.path was being incorrectly used to detect if the author had set up a hyperlink for the tile's title. The correct value is content.href (content.path returns a path to a page the returns the html for the component on a page by itself... kinda handy but not for this).

Changes the componentGroup for the main-navigation-oracle-give component from Admin to .hidden to remove the component from sidekick.